Fix time to complete level calculation.
The game has a frame rate of 60fps, i.e. 16.67ms per frame. The
original code was counting 6000 hundredths of a second (i.e. 60000ms)
per frame, making the elapsed time calculation far too long.

In order to keep all the rest of the code working in hundredths of a
second, this patch uses a floating point number to 'accurately' count
the time and then shifts whole units of hundredths of a second into
the variable 'ten_mili_seconds' which is used by the rest of the code.